Move the DTOs into an inmanta.dto package of thematic modules
inmanta/data/model.py held all 81 data transfer objects in one module, inside the
inmanta.data package whose __init__ is the database access layer. Anything that
needed a single DTO therefore executed the whole database layer, and every new
DTO made that one module heavier.

The DTOs now live in inmanta/dto/, split by subject: status, compile,
environment, log, resource, diff, paging, parameter, agent, desiredstate, dryrun,
notification, auth, discovery, pip, scheduler and code. inmanta/dto/__init__.py
is deliberately empty, because re-exporting from it would recreate exactly the
problem this solves: importing one theme would load them all.

That confines each heavy dependency to the theme that needs it. inmanta.ast.export
is only reachable through dto.compile, inmanta.deploy.state and asyncpg only
through dto.resource. Needing PipConfig costs 850 loaded modules through
inmanta.dto.pip against 1547 through inmanta.data.model, with the AST exporter,
the deploy state machine, asyncpg and the database layer all absent.

inmanta/data/model.py becomes a re-export of all 81 names, so no call site
changes and the stable API contract holds. It is the one place that still loads
everything, which makes it the thing to delete once call sites have moved.

ResourceLog is defined next to LogLine rather than with the other resource DTOs.
The shim, inmanta.data and inmanta.resources form one strongly connected
component for mypy, and inside it the pydantic plugin does not always see a base
model that lives in another module, so it synthesised ResourceLog.__init__
without the inherited fields. Keeping a base and its only subclass together
avoids relying on that ordering.
