+++ src/inmanta/dto/__init__.py
@@ -1,0 +1,20 @@
+"""
+Copyright 2019 Inmanta
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+Contact: code@inmanta.com
+"""
+
+# Deliberately empty. The data transfer objects live in thematic submodules so that importing the ones you need does
+# not pull in the dependencies of all the others. Import from inmanta.dto.<theme>, never from this package.
