+++ src/inmanta/dto/dryrun.py
@@ -1,0 +1,38 @@
+"""
+Copyright 2019 Inmanta
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+Contact: code@inmanta.com
+"""
+
+import datetime
+import uuid
+from typing import Optional
+
+from inmanta.dto.diff import ResourceDiff
+from inmanta.types import BaseModel as BaseModel
+
+
+class DryRun(BaseModel):
+    id: uuid.UUID
+    environment: uuid.UUID
+    model: int
+    date: Optional[datetime.datetime] = None
+    total: int = 0
+    todo: int = 0
+
+
+class DryRunReport(BaseModel):
+    summary: DryRun
+    diff: list[ResourceDiff]
